Revert "Revert "Merge branch 'main' into misc""
This reverts commit 97321564609b12b1363b6af097926d303319efcf.

+++ proyecto_musica/authentication/serializers.py
@@ -1,25 +1,46 @@
 from rest_framework import serializers
-from authentication.models import User, Skills, User_Skills, Genres, User_Genres
+from authentication.models import User, Skills, User_Skills, Genres, User_Genres, User_Artists, Genders
+from api.models import Images
 from authentication.functions import List_Fields, get_list_field
 from rest_framework.validators import UniqueValidator
 from django.contrib.auth.models import BaseUserManager, AbstractBaseUser
 
 class RegisterSerializer(serializers.ModelSerializer):
     password = serializers.CharField(max_length=128, min_length=6, write_only=True)
+    gender_name = serializers.SerializerMethodField()
     skills = serializers.SerializerMethodField()
     genres = serializers.SerializerMethodField()
+    artists = serializers.SerializerMethodField()
+    pictures = serializers.SerializerMethodField()
 
     class Meta():
         model=User
-        fields=('username','email','password','skills','genres')
+        fields=('username','email','first_name','last_name', 'gender', 
+                'gender_name','DOB','about_me', 'password','skills','genres',
+                'artists','pictures')
 
+    def get_gender_name(self, obj):
+        gender_id = obj.gender_id
+        if gender_id:
+            res = Genders.objects.get(gender_id=gender_id)
+            return res.gender_name
+    
     def get_skills(self, obj):
         skills = self.context.get("skills")        
-        return get_list_field(None, obj.email, "skill", skills)
+        return get_list_field(obj.id, "skill", skills)
     
     def get_genres(self, obj):
         genres = self.context.get("genres")        
-        return get_list_field(None, obj.email, "genre", genres)
+        return get_list_field(obj.id, "genre", genres)
+
+    def get_artists(self, obj):
+        artists = self.context.get("artists")
+        return get_list_field(obj.id, "artist", artists)
+ 
+    def get_pictures(self, obj):
+        query = Images.objects.filter(user_id=obj.id).values('image_id', 
+                    'url', 'title')
+        return list(query) if query else None
 
     def create(self, validated_data):
         user =  User.objects.create_user(**validated_data)
@@ -35,6 +56,9 @@ class RegisterSerializer(serializers.ModelSerializer):
                 elif field_name == 'genres':
                     for obj in field_list:
                         User_Genres.objects.create(user_id=user_id, genre_id=obj)
+                elif field_name == 'artists':
+                    for obj in field_list:
+                        User_Artists.objects.create(user_id=user_id, artist=obj)
 
         return user
 
@@ -45,20 +69,32 @@ class EditSerializer(serializers.ModelSerializer):
     username = serializers.CharField(validators=[UniqueValidator(queryset=User.objects.all())]) 
     skills = serializers.SerializerMethodField()
     genres = serializers.SerializerMethodField()
+    artists = serializers.SerializerMethodField()
+    gender_name = serializers.SerializerMethodField()
 
     class Meta:
         model=User
-        fields=('username','email','password','skills','genres')
+        fields=('username','email','first_name','last_name','gender',
+                'gender_name','DOB','about_me','password','skills','genres',
+                'artists')
 
+    def get_gender_name(self, obj):
+        gender_id = obj.gender_id
+        if gender_id:
+            res = Genders.objects.get(gender_id=gender_id)
+            return res.gender_name
+    
     def get_skills(self, obj):
-        id = self.context.get("id")
         skills = self.context.get("skills")
-        return get_list_field(id, None, "skill", skills)
+        return get_list_field(obj.id, "skill", skills)
 
     def get_genres(self, obj):
-        id = self.context.get("id")
         genres = self.context.get("genres")        
-        return get_list_field(id, None, "genre", genres)
+        return get_list_field(obj.id, "genre", genres)
+    
+    def get_artists(self, obj):
+        artists = self.context.get("artists")
+        return get_list_field(obj.id, "artist", artists)
     
     def update(self, instance, validated_data):
         original_email = validated_data.get('email', instance.email)
@@ -67,25 +103,35 @@ class EditSerializer(serializers.ModelSerializer):
 
         instance.email = BaseUserManager.normalize_email(original_email)
         instance.username = AbstractBaseUser.normalize_username(original_username)
+        instance.first_name = validated_data.get('first_name', instance.first_name)
+        instance.last_name = validated_data.get('last_name', instance.last_name)
+        instance.gender = validated_data.get('gender', instance.gender)
+        instance.DOB = validated_data.get('DOB', instance.DOB)
+        instance.about_me = validated_data.get('about_me', instance.about_me)
         if 'password' in validated_data:
             password = validated_data.pop('password')
             instance.set_password(original_password)
         
         instance.save()
         
-        # update skills/genres in corresponding tables
-        id = self.context.get('id')
+        # update other fields in corresponding tables
+        id = instance.id
         for field in List_Fields:
             field_name = field.value
             field_list = self.context.get(field_name)
-            if field_name == 'skills':
-                User_Skills.objects.filter(user_id=id).delete()
-                for obj in field_list:    
-                    User_Skills.objects.create(user_id=id, skill_id=obj)
-            elif field_name == 'genres':
-                User_Genres.objects.filter(user_id=id).delete()
-                for obj in field_list:    
-                    User_Genres.objects.create(user_id=id, genre_id=obj)
+            if field_list is not None:
+                if field_name == 'skills':
+                    User_Skills.objects.filter(user_id=id).delete()
+                    for obj in field_list:    
+                        User_Skills.objects.create(user_id=id, skill_id=obj)
+                elif field_name == 'genres':
+                    User_Genres.objects.filter(user_id=id).delete()
+                    for obj in field_list:    
+                        User_Genres.objects.create(user_id=id, genre_id=obj)
+                elif field_name == 'artists':
+                    User_Artists.objects.filter(user_id=id).delete()
+                    for obj in field_list:    
+                        User_Artists.objects.create(user_id=id, artist=obj)
         
         return instance
